requirements.txt
# ═══════════════════════════════════════════════════════════════════════
# THE FORGE — AI Build Engine
# requirements.txt
# ═══════════════════════════════════════════════════════════════════════

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Redis & job queue
redis==5.2.1
rq==2.0.0

# Scheduling
APScheduler==3.10.4

# AI clients
anthropic==0.40.0
openai==1.58.1
tavily-python==0.5.0

# Notifications
python-telegram-bot==21.9

# Document parsing
python-docx==1.1.2
pypdf==5.1.0

# Logging
loguru==0.7.3

# HTTP
httpx==0.28.1
aiohttp==3.11.11

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Content processing
feedparser==6.0.11
tiktoken==0.8.0
beautifulsoup4==4.12.3

# Utilities
aiofiles==24.1.0

# Monitoring
sentry-sdk[fastapi]==2.19.2

# Code quality — runs on generated output before packaging
black==24.10.0
isort==5.13.2

# Security scanning — runs on generated output before packaging
pip-audit==2.7.3
detect-secrets==1.5.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
